修改招聘，商务合同

+++ modules/businessproject/businessproject-api/src/main/java/com/bjike/goddess/businessproject/excel/BusinessContractExport.java
@@ -197,7 +197,7 @@ public class BusinessContractExport extends BaseTO {
      * 合同规模数量
      */
     @ExcelHeader(name="合同规模数量")
-    private Double scaleContract;
+    private Integer scaleContract;
 
     /**
      * 合同规模数是否有差异
@@ -704,11 +704,11 @@ public class BusinessContractExport extends BaseTO {
         this.scale = scale;
     }
 
-    public Double getScaleContract() {
+    public Integer getScaleContract() {
         return scaleContract;
     }
 
-    public void setScaleContract(Double scaleContract) {
+    public void setScaleContract(Integer scaleContract) {
         this.scaleContract = scaleContract;
     }
